Base de datos de 10 archivos ...

+++ main.rb
@@ -41,13 +41,14 @@ end
 post '/save' do
   pp params
   name = params[:fname]
-  c  = PL0Program.first(:name => name)
-  puts "prog <#{c.inspect}>"
-  if c
-    c.source = params["input"]
-    c.save
-  else
-    if PL0Program.all.size > 9
+  if name != "test"
+    c  = PL0Program.first(:name => name)
+    puts "prog <#{c.inspect}>"
+    if c
+      c.source = params["input"]
+      c.save
+    else
+      if PL0Program.all.size > 9
         c = PL0Program.all.sample
         c.destroy
       end
